+++ packages/amazon-cognito-identity-js/src/Platform/index.js
@@ -7,26 +7,10 @@ import { version } from './version';
 const BASE_USER_AGENT = `aws-amplify/${version}`;
 
 export const Platform = {
-	userAgent: `${BASE_USER_AGENT} js`,
-	product: '',
-	navigator: null,
-	isReactNative: false,
+	userAgent: BASE_USER_AGENT,
+	isReactNative:
+		typeof navigator !== 'undefined' && navigator.product === 'ReactNative',
 };
-
-if (typeof navigator !== 'undefined' && navigator.product) {
-	Platform.product = navigator.product || '';
-	Platform.navigator = navigator || null;
-	switch (navigator.product) {
-		case 'ReactNative':
-			Platform.userAgent = `${BASE_USER_AGENT} react-native`;
-			Platform.isReactNative = true;
-			break;
-		default:
-			Platform.userAgent = `${BASE_USER_AGENT} js`;
-			Platform.isReactNative = false;
-			break;
-	}
-}
 
 export const getUserAgent = () => {
 	return Platform.userAgent;