+++ packages/storage/src/types/AWSS3Provider.ts
@@ -5,6 +5,7 @@ import {
 	CopyObjectRequest,
 	_Object,
 	DeleteObjectCommandOutput,
+	HeadObjectRequest,
 } from '@aws-sdk/client-s3';
 import { StorageOptions, StorageAccessLevel } from './Storage';
 import {
@@ -51,6 +52,12 @@ export type S3ProviderGetConfig = CommonStorageOptions & {
 	validateObjectExistence?: boolean;
 };
 
+export type S3ProviderGetPropertiesConfig = CommonStorageOptions & {
+	SSECustomerAlgorithm?: HeadObjectRequest['SSECustomerAlgorithm'];
+	SSECustomerKey?: HeadObjectRequest['SSECustomerKey'];
+	SSECustomerKeyMD5?: HeadObjectRequest['SSECustomerKeyMD5'];
+};
+
 export type S3ProviderGetOuput<T> = T extends { download: true }
 	? GetObjectCommandOutput
 	: string;
@@ -84,6 +91,15 @@ export type ResumableUploadConfig = {
 	errorCallback?: (err: any) => any;
 };
 
+/**
+ * Configuration options for the S3 put function.
+ *
+ * @remarks
+ * The acl parameter may now only be used for S3 buckets with specific Object Owner settings.
+ * Usage of this parameter is not considered a recommended practice:
+ *  {@link https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html}
+ *
+ */
 export type S3ProviderPutConfig = CommonStorageOptions &
 	(
 		| _S3ProviderPutConfig
@@ -133,6 +149,15 @@ export type S3CopySource = S3CopyTarget;
 
 export type S3CopyDestination = Omit<S3CopyTarget, 'identityId'>;
 
+/**
+ * Configuration options for the S3 copy function.
+ *
+ * @remarks
+ * The acl parameter may now only be used for S3 buckets with specific Object Owner settings.
+ * Usage of this parameter is not considered a recommended practice:
+ *  {@link https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html}
+ *
+ */
 export type S3ProviderCopyConfig = Omit<CommonStorageOptions, 'level'> & {
 	provider?: 'AWSS3';
 	bucket?: CopyObjectRequest['Bucket'];
@@ -153,6 +178,14 @@ export type S3ProviderCopyConfig = Omit<CommonStorageOptions, 'level'> & {
 
 export type S3ProviderCopyOutput = {
 	key: string;
+};
+
+export type S3ProviderGetPropertiesOutput = {
+	contentType: string;
+	contentLength: number;
+	eTag: string;
+	lastModified: Date;
+	metadata: Record<string, string>;
 };
 
 export type PutResult = {