+++ packages/api/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.3.0...@aws-amplify/api@5.3.1) (2023-06-21)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+# [5.3.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.2.1...@aws-amplify/api@5.3.0) (2023-06-20)
+
+### Features
+
+- **core:** API detection for 8 framework targets ([#11384](https://github.com/aws-amplify/amplify-js/issues/11384)) ([388f207](https://github.com/aws-amplify/amplify-js/commit/388f2074db0640f2d22aa7cd1a44d8eb8f2301d2))
+- Custom user agent enhancement for api-rest ([#11457](https://github.com/aws-amplify/amplify-js/issues/11457)) ([4315ba0](https://github.com/aws-amplify/amplify-js/commit/4315ba05ad2e4463a4ef19431f20f8d5308cdc62))
+- user agent enhancements - cross-category api>api-graphql>pubsub ([#11368](https://github.com/aws-amplify/amplify-js/issues/11368)) ([a88dda2](https://github.com/aws-amplify/amplify-js/commit/a88dda202919aa1d4dc2af48f33c3533d8626261))
+
+## [5.2.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.2.0...@aws-amplify/api@5.2.1) (2023-06-14)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+# [5.2.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.1.1...@aws-amplify/api@5.2.0) (2023-06-05)
+
+### Bug Fixes
+
+- address feedbacks ([2d6eecf](https://github.com/aws-amplify/amplify-js/commit/2d6eecfa4763a6cfb6aeaabedd49a530c6420dcd))
+
+### Features
+
+- **clients:** cognito identity client ([#11213](https://github.com/aws-amplify/amplify-js/issues/11213)) ([67e4017](https://github.com/aws-amplify/amplify-js/commit/67e40171385f02d0c9448fdc3e036d63e009ea34))
+- **clients:** support CN partition by adding DNS suffix resolver ([#11311](https://github.com/aws-amplify/amplify-js/issues/11311)) ([9de2975](https://github.com/aws-amplify/amplify-js/commit/9de297519fdbaaf1e9b4ae98f12aed4137400222))
+
+## [5.1.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.1.0...@aws-amplify/api@5.1.1) (2023-05-27)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+# [5.1.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.31...@aws-amplify/api@5.1.0) (2023-05-12)
+
+### Features
+
+- **core:** refactor signer implementation to remove polyfill ([#11319](https://github.com/aws-amplify/amplify-js/issues/11319)) ([1603af3](https://github.com/aws-amplify/amplify-js/commit/1603af38ea7faa163f7fcb92fd09c056454c8a0a)), closes [#11273](https://github.com/aws-amplify/amplify-js/issues/11273) [#11307](https://github.com/aws-amplify/amplify-js/issues/11307) [#11310](https://github.com/aws-amplify/amplify-js/issues/11310) [#11120](https://github.com/aws-amplify/amplify-js/issues/11120) [#11320](https://github.com/aws-amplify/amplify-js/issues/11320)
+
+## [5.0.31](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.30...@aws-amplify/api@5.0.31) (2023-05-04)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## [5.0.30](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.29...@aws-amplify/api@5.0.30) (2023-04-27)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## [5.0.29](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.28...@aws-amplify/api@5.0.29) (2023-04-20)
+
+**Note:** Version bump only for package @aws-amplify/api
+
 ## [5.0.28](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.27...@aws-amplify/api@5.0.28) (2023-04-18)
 
 **Note:** Version bump only for package @aws-amplify/api